testing suite now includes several regression tests along with a full transcoding workflow for easier validation

+++ source/example-utils.hpp
@@ -70,10 +70,10 @@ struct memory_stream : virtual memory_buffer, public std::istream
 class manual_timer
 {
     std::chrono::high_resolution_clock::time_point t0;
-    double timestamp{ 0.f };
+    double timestamp{ 0.0 };
 public:
     void start() { t0 = std::chrono::high_resolution_clock::now(); }
-    void stop() { timestamp = std::chrono::duration<float>(std::chrono::high_resolution_clock::now() - t0).count() * 1000; }
+    void stop() { timestamp = std::chrono::duration<double>(std::chrono::high_resolution_clock::now() - t0).count() * 1000.0; }
     const double & get() { return timestamp; }
 };
